Tests updated and moved
