+++ Src/Dft.DTRO.Tests/CodeiumTests/Tra/Controller/TraControllerCreateFromBodyTests.cs
@@ -1,0 +1,67 @@
+﻿using DfT.DTRO.Models.SwaCode;
+
+namespace Dft.DTRO.Tests.CodeiumTests.Tra.Controller;
+
+public class TraControllerCreateFromBodyTests
+{
+    private readonly Mock<ITraService> _traServiceMock;
+    private readonly Mock<ILogger<TraController>> _loggerMock;
+    private readonly TraController _controller;
+
+    public TraControllerCreateFromBodyTests()
+    {
+        _traServiceMock = new Mock<ITraService>();
+        _loggerMock = new Mock<ILogger<TraController>>();
+        _controller = new TraController(_traServiceMock.Object, _loggerMock.Object);
+    }
+
+
+    [Fact]
+    public async Task CreateFromBody_ReturnsCreated_WithGuidResponse()
+    {
+        // Arrange
+        var request = new SwaCodeRequest { TraId = 1 };
+        var response = new GuidResponse { Id = Guid.NewGuid() };
+        _traServiceMock.Setup(service => service.SaveTraAsync(request)).ReturnsAsync(response);
+
+        // Act
+        var result = await _controller.CreateFromBody(request);
+
+        // Assert
+        var createdResult = Assert.IsType<CreatedAtActionResult>(result);
+        Assert.Equal(201, createdResult.StatusCode);
+        Assert.Equal(response, createdResult.Value);
+    }
+
+    [Fact]
+    public async Task CreateFromBody_ReturnsBadRequest_WhenInvalidOperationExceptionIsThrown()
+    {
+        // Arrange
+        var request = new SwaCodeRequest { TraId = 1 };
+        _traServiceMock.Setup(service => service.SaveTraAsync(request)).ThrowsAsync(new InvalidOperationException("Invalid operation"));
+
+        // Act
+        var result = await _controller.CreateFromBody(request);
+
+        // Assert
+        var badRequestResult = Assert.IsType<BadRequestObjectResult>(result);
+        Assert.Equal(400, badRequestResult.StatusCode);
+        Assert.IsType<ApiErrorResponse>(badRequestResult.Value);
+    }
+
+    [Fact]
+    public async Task CreateFromBody_ReturnsInternalServerError_WhenExceptionIsThrown()
+    {
+        // Arrange
+        var request = new SwaCodeRequest { TraId = 1 };
+        _traServiceMock.Setup(service => service.SaveTraAsync(request)).ThrowsAsync(new Exception());
+
+        // Act
+        var result = await _controller.CreateFromBody(request);
+
+        // Assert
+        var statusCodeResult = Assert.IsType<ObjectResult>(result);
+        Assert.Equal(500, statusCodeResult.StatusCode);
+        Assert.IsType<ApiErrorResponse>(statusCodeResult.Value);
+    }
+}
